src/model-service/pyproject.toml
[project]
name = "tsm_mach_le_data_model-service"
version = "0.1.0"
description = "A GNN-based link prediction/regression model for movie recommendation."
readme = "README.MD"
requires-python = ">=3.12,<3.13"
dependencies = [
  "pandas",
  "wandb==0.19.8",
  "python-dotenv",
  "onnx",
  "onnxruntime",
  "fastapi",
  "uvicorn",
]

[project.optional-dependencies]
cpu = [
  "torch==2.8.0",
  "torchvision",
  "torch-geometric",
  "torch-scatter",
  "torch-sparse",
  "torch-cluster",
  "torch-spline-conv",
]

gpu = [
  "torch==2.8.0",
  "torchvision",
  "torch-geometric",
  "torch-scatter",
  "torch-sparse",
  "torch-cluster",
  "torch-spline-conv",
]


[tool.uv]
environments = [
  "python_full_version >= '3.12' and python_full_version < '3.13'",
]
conflicts = [[{ extra = "cpu" }, { extra = "gpu" }]]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pyg-cpu"
url = "https://data.pyg.org/whl/torch-2.8.0+cpu.html"
format = "flat"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[[tool.uv.index]]
name = "pyg-cu129"
url = "https://data.pyg.org/whl/torch-2.8.0+cu129.html"
format = "flat"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu129", extra = "gpu" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu129", extra = "gpu" },
]

torch-scatter = [
  { index = "pyg-cpu", extra = "cpu" },
  { index = "pyg-cu129", extra = "gpu" },
]
torch-sparse = [
  { index = "pyg-cpu", extra = "cpu" },
  { index = "pyg-cu129", extra = "gpu" },
]
torch-cluster = [
  { index = "pyg-cpu", extra = "cpu" },
  { index = "pyg-cu129", extra = "gpu" },
]
torch-spline-conv = [
  { index = "pyg-cpu", extra = "cpu" },
  { index = "pyg-cu129", extra = "gpu" },
]

[tool.uv.extra-build-dependencies]
torch-cluster = [{ requirement = "torch", match-runtime = true }]

[tool.fastapi]
entrypoint = "main.py"
